ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pytesseract==0.3.13
Pillow==11.1.0
spacy==3.8.4
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
